+++ log-guard-core/src/main/java/io/github/dancan254/logguard/Pii.java
@@ -8,7 +8,7 @@ import java.lang.annotation.Target;
 
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD, ElementType.RECORD_COMPONENT, ElementType.METHOD})
+@Target({ElementType.FIELD, ElementType.RECORD_COMPONENT})
 public @interface Pii {
 
     MaskStrategy strategy() default MaskStrategy.REDACT;